Refactor Petrinaut optimization handler
